Api Summoner Id GetMapping Error fix

+++ src/main/java/org/cnu/realcoding/lolapi/team13/lolapiteam13/domain/SummonerId.java
@@ -7,8 +7,8 @@ public class SummonerId {
     private int profileIconId;
     private String name;
     private String puuid;
-    private int summonerLevel;
+    private long summonerLevel;
     private String accountId;
     private String id;
-    private int revisionDate;
+    private long revisionDate;
 }
